% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{activation_box_sum_sigmoid}
\alias{activation_box_sum_sigmoid}
\title{Box Constraints and Sum Constraints Activation with Sigmoid Scaling}
\usage{
activation_box_sum_sigmoid(min_weight = -1, max_weight = 1, target_sum = 0)
}
\arguments{
\item{min_weight}{Minimum weight to map to (default: -1.0).}

\item{max_weight}{Maximum weight to map to (default: 1.0).}

\item{target_sum}{The target sum of the weights (default: 0.0).}
}
\value{
A Keras activation function that maps the input to the desired range using the sigmoid function and normalizes the sum.
}
\description{
This activation function first maps the output using a sigmoid function to ensure that the predicted weights fall
within a specified range (\code{min_weight}, \code{max_weight}), then normalizes the weights to ensure they sum to \code{target_sum}.
}
\examples{
\dontrun{
data(data_ml)
test_data_ml <- data_ml \%>\% filter(stock_id <= 5)
train_data_ex <- test_data_ml |> filter(date<="2012-12-31") |> arrange(date, stock_id)
test_data_ex <- test_data_ml |> filter(date>"2012-12-31") |> arrange(date, stock_id)
config_box_sum_sigmoid <- list(
  layers = list(
    list(type = "dense", units = 32, activation = "relu"),
    list(type = "dense", units = 16, activation = "relu"),
    list(type = "dense", units = 1, activation = activation_box_sum_sigmoid(min_weight=0.5,max_weight=1,target_sum=1))
  ),
  loss = "dummy_mse_loss",
  optimizer = list(name = "optimizer_adam", learning_rate = 0.001),
  epochs = 10,
  batch_size = 128,
  verbose = 1,
  seeds = c(42)
)
weights <- keras_weights(train_data_ex, test_data_ex, config_box_sum_sigmoid)
print(weights)
}
}
