% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{activation_sum_sigmoid}
\alias{activation_sum_sigmoid}
\title{Sum Constraints Activation with Sigmoid Scaling}
\usage{
activation_sum_sigmoid(target_sum = 0)
}
\arguments{
\item{target_sum}{The target sum of the weights (default: 0.0).}

\item{min_weight}{Minimum weight (default: -1.0).}

\item{max_weight}{Maximum weight (default: 1.0).}
}
\value{
A Keras activation function that normalizes the weights to meet the target sum.
}
\description{
This activation function normalizes the predicted weights to ensure they sum to \code{target_sum} after applying the sigmoid function.
}
\examples{
\dontrun{
data(data_ml)
test_data_ml <- data_ml \%>\% filter(stock_id <= 5)
train_data_ex <- test_data_ml |> filter(date<="2012-12-31") |> arrange(date, stock_id)
test_data_ex <- test_data_ml |> filter(date>"2012-12-31") |> arrange(date, stock_id)
config_sum_sigmoid <- list(
  layers = list(
    list(type = "dense", units = 32, activation = "relu"),
    list(type = "dense", units = 16, activation = "relu"),
    list(type = "dense", units = 5, activation = activation_sum_sigmoid(target_sum = 0))
  ),
  loss = "dummy_mse_loss",
  optimizer = list(name = "optimizer_adam", learning_rate = 0.001),
  epochs = 10,
  batch_size = 128,
  verbose = 1,
  seeds = c(42)
)
weights <- keras_weights(train_data_ex, test_data_ex, config_sum_sigmoid)
print(weights)
}
}
