% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{add_extra_data}
\alias{add_extra_data}
\title{Add Extra Data to returnPrediction S3 Object}
\usage{
add_extra_data(
  return_prediction_object,
  extra_data,
  new_data_name = "extra_data"
)
}
\arguments{
\item{return_prediction_object}{An object of class \code{returnPrediction}.}

\item{extra_data}{A tibble or data.frame containing additional data to be added. Must include \code{stock_id} and \code{date} columns and match the dimensions of the existing data.}

\item{new_data_name}{A character string specifying the name under which the extra data will be stored within the \code{returnPrediction} object.}
}
\value{
An updated \code{returnPrediction} object with the added extra data.
}
\description{
This function allows users to add supplementary datasets to an existing \code{returnPrediction} object. The extra data can include any relevant information, but must have the same \code{stock_id} and \code{date} combinations as the existing data in the object.
}
\examples{
\dontrun{
# Initialize the returnPrediction object
data <- tibble::tibble(
  stock_id = 1:100,
  date = seq.Date(Sys.Date(), by = "day", length.out = 100),
  return_label = runif(100)
)
# Create the returnPrediction object
rp <- create_return_prediction(data, "return_label")

# Add extra data regarding weights.
extra_data <- tibble::tibble(
  stock_id = 1:100,
   date = seq.Date(Sys.Date(), by = "day", length.out = 100),
  sector = sample(c("Technology", "Healthcare", "Finance"), 100, replace = TRUE),
  country = sample(c("USA", "Germany", "Japan"), 100, replace = TRUE)
)
rp <- add_extra_data(rp, extra_data, new_data_name = "sector_info")

# View the added extra data
print(rp$extra_data$sector_info)
}

}
