% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{add_model_prediction}
\alias{add_model_prediction}
\title{Add model prediction to returnPrediction S3 object}
\usage{
add_model_prediction(
  return_prediction,
  model_function,
  config,
  new_predictions
)
}
\arguments{
\item{return_prediction}{Add model prediction to this object}

\item{model_function}{Name of the prediction (model) function}

\item{config}{Configuration for the model}

\item{new_predictions}{Tibble with stock_id, date, and prediction columns}
}
\value{
returnPrediction S3 object with added model prediction
}
\description{
Add model prediction to returnPrediction S3 object
}
\examples{
\dontrun{
# Initialize the returnPrediction object
data <- tibble::tibble(
  stock_id = 1:100,
  date = seq.Date(Sys.Date(), by = "day", length.out = 100),
  return_label = runif(100)
)
# Create the returnPrediction object
rp <- create_return_prediction(data, "return_label")
# Sample new prediction data
new_predictions <- tibble::tibble(
  stock_id = 1:100,
  date = seq.Date(Sys.Date(), by = "day", length.out = 100),
  prediction = runif(100)  # Example prediction values
  )
# Model function and config for demonstration
model_function <- "xgb_pred"  # Placeholder for your model function
config <- list(max_depth = 4, eta = 0.1)  # Example configuration
# Adding the new model predictions
rp <- add_model_prediction(rp, model_function, config, new_predictions)
print(rp$predictions)
print(rp$errors)
print(rp$models)
}
}
