% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{add_weight_model}
\alias{add_weight_model}
\title{Add weight model and predictions to portfolioReturns S3 object}
\usage{
add_weight_model(portfolio_object, model_name, weight_config, new_weights)
}
\arguments{
\item{portfolio_object}{A portfolioReturns object.}

\item{model_name}{Name of the weight prediction model.}

\item{weight_config}{Configuration for the model.}

\item{new_weights}{Data frame containing stock_id, date, and the new weights.}
}
\value{
Updated portfolioReturns S3 object with added weight model and performance data.
}
\description{
Add weight model and predictions to portfolioReturns S3 object
}
\examples{
\dontrun{
data <- tibble::tibble(
  stock_id = 1:100,
  date = rep(Sys.Date(), 100),
  weights = runif(100)
)
pf <- create_portfolios(data, "return_label")
pf <- add_weight_model(pf, "keras_weights", list(epochs = 50), data)
print(pf)
}
}
