% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting_returns.R
\name{backtesting_returns}
\alias{backtesting_returns}
\title{Backtesting Returns Function}
\usage{
backtesting_returns(
  data,
  return_prediction_object = NULL,
  return_label,
  features,
  rolling = TRUE,
  window_size,
  step_size = "1 month",
  offset = "0",
  in_sample = TRUE,
  ml_config,
  append = FALSE,
  num_cores = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{ML dataset (tibble/data.frame) in long format containing the features, return label, stock IDs (first column), and dates (second column).
The dataset should be free of missing values and, if necessary, balanced in terms of the number of stocks available at each point in time.}

\item{return_prediction_object}{An object of class \code{returnPrediction} to store the predictions. Default is \code{NULL}, which creates a new object.
If an existing object is provided and \code{append = TRUE}, new predictions are added. If \code{append = FALSE}, the object is overwritten.}

\item{return_label}{The prediction label to be used for the ML model. It should be appropriately shifted (e.g., the label for date \code{t} should correspond to returns from date \code{t+1}).}

\item{features}{A character vector of feature names to be used for the ML model.}

\item{rolling}{Logical indicating whether to use a rolling window approach (\code{TRUE}) or an expanding window approach (\code{FALSE}) for prediction.}

\item{window_size}{The size of the window for the rolling window approach. Can be specified in number of time steps or as a period (e.g., "1 year", "6 months").
If \code{rolling = FALSE}, this represents the starting window size for the expanding window approach.}

\item{step_size}{The interval at which the prediction window moves forward (e.g., "1 month", "15 days"). Default is "1 month".}

\item{offset}{The size of the data to be left unused between training and prediction to avoid look-ahead bias. Can be specified in time steps or periods (e.g., "1 month"). Default is "0".}

\item{in_sample}{Logical indicating whether to provide in-sample predictions for the training period (+ offset). Default is \code{TRUE}.}

\item{ml_config}{A list containing configurations for the ML models. Each element should specify the prediction function and its respective configurations.}

\item{append}{Logical indicating whether to append the predicted returns to the original dataset (\code{TRUE}) or overwrite the existing \code{returnPrediction} object (\code{FALSE}).}

\item{num_cores}{The number of cores to be used for parallel processing. If set to \code{NULL}, ML iterations are performed sequentially.}

\item{verbose}{Logical indicating whether to display detailed progress messages. Default is \code{FALSE}.}
}
\value{
An S3 object of class \code{returnPrediction} containing models, predictions, actual returns, and errors.
}
\description{
This function utilizes machine learning models to predict asset returns, which are subsequently used for portfolio trading strategies.
}
\examples{
\dontrun{
data(data_ml)
data_subset <- data_ml |>  filter(stock_id<=30)
return_label <- "R1M_Usd"
features <- c("Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")
rolling <- TRUE
window_size <- "5 years"
step_size <- "3 months"
offset <- "1 year"
in_sample <- TRUE
ml_config <- list(
  ols_pred = list(pred_func = "ols_pred", config = list()),
  xgb_pred = list(
    pred_func = "xgb_pred",
    config1 = list(nrounds = 100, max_depth = 3, eta = 0.3, objective = "reg:squarederror"),
    config2 = list(nrounds = 100, max_depth = 4, eta = 0.1, objective = "reg:squarederror")
  )
)
rp <- backtesting_returns(
  data = data_subset,
  return_prediction_object = NULL,
  return_label = return_label,
  features = features,
  rolling = rolling,
  window_size = window_size,
  step_size = step_size,
  offset = offset,
  in_sample = in_sample,
  ml_config = ml_config,
  append = FALSE,
  num_cores = NULL,
  verbose = TRUE
)
}
}
