% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_active_share}
\alias{calculate_active_share}
\title{Calculate Active Share}
\usage{
calculate_active_share(weights, benchmark_weights)
}
\arguments{
\item{weights}{Portfolio weights data frame.}

\item{benchmark_weights}{Benchmark weights data frame.}
}
\value{
Active share value per portfolio per date.
}
\description{
Calculate Active Share
}
