% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{calculate_errors}
\alias{calculate_errors}
\title{Calculate errors between predictions and actual returns}
\usage{
calculate_errors(predictions, actual_returns)
}
\arguments{
\item{predictions}{Predictions to be evaluated}

\item{actual_returns}{Actual returns}
}
\value{
List with Mean Squared Error (MSE), Root Mean Squared Error (RMSE), Mean Absolute Error (MAE), and Hit Ratio
}
\description{
Calculate errors between predictions and actual returns
}
