% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_turnover}
\alias{calculate_turnover}
\title{Calculate Turnover}
\usage{
calculate_turnover(weights, actual_returns)
}
\arguments{
\item{weights}{Portfolio weights data frame.}

\item{actual_returns}{Portfolio returns data frame.}
}
\value{
Per-period turnover values and the total turnover across all periods.
}
\description{
Calculate Turnover
}
