% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{caret_wrapper}
\alias{caret_wrapper}
\title{Wrapper for Model Training using Caret}
\usage{
caret_wrapper(train_data, test_data, config = list(), seed = NULL)
}
\arguments{
\item{train_data}{A data frame with stock_id, date, return_label, and features.}

\item{test_data}{A data frame with stock_id, date, and features.}

\item{config}{A list of configuration settings, containing the method, tuneGrid, and trControl for caret.}

\item{seed}{integer. Seed for random number generation}
}
\value{
A tibble with stock_id, date, and predicted returns for the test data.
}
\description{
Trains machine learning models using caret, supporting various models (e.g., Elastic Net, SVM, etc.)
with optional hyperparameter tuning and cross-validation.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c(1,2,96,5:10)]
test_data_ex <- data_ml[101:150, c(1,2,5:10)]
config <- list(method = "glmnet", tuneGrid = expand.grid(alpha = 0.5, lambda = 0.1),
               trControl = caret::trainControl(method = "cv", number = 5))
caret_wrapper(train_data_ex, test_data_ex, config)
}
}
