% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_period_to_days}
\alias{convert_period_to_days}
\title{Helper function to convert any time period to days}
\usage{
convert_period_to_days(period)
}
\arguments{
\item{period}{A string that contains a time period, e.g. "3 days", "1 week", "2 months", "1 year"}
}
\value{
The number of days in the time period
}
\description{
Helper function to convert any time period to days
}
\examples{
convert_period_to_days("3 days")
convert_period_to_days("1 week")
convert_period_to_days("2 months")
convert_period_to_days("1 year")
}
