% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{create_portfolios}
\alias{create_portfolios}
\title{Create portfolioReturns S3 object}
\usage{
create_portfolios(data, label)
}
\arguments{
\item{data}{Data frame containing stock_id, date, and the specified label column}

\item{label}{Column name of the label column}
}
\value{
A portfolioReturns S3 object
}
\description{
Create portfolioReturns S3 object
}
\examples{
data <- tibble(
  stock_id = 1:100,
  date = c(rep(Sys.Date(), 50), rep(Sys.Date() + 1, 50)),
  return_label = runif(100)
)
# Create the returnPrediction object
pf <- create_portfolios(data, "return_label")
# Check the initial setup
print(pf)
}
