% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{create_return_prediction}
\alias{create_return_prediction}
\title{Create returnPrediction S3 object}
\usage{
create_return_prediction(data, label)
}
\arguments{
\item{data}{Data that will be used for prediction}

\item{label}{Column name of the actual return}
}
\value{
returnPrediction S3 object
}
\description{
Create returnPrediction S3 object
}
\examples{
# Initialize the returnPrediction object
data <- tibble(
  stock_id = 1:100,
  date = seq.Date(Sys.Date(), by = "day", length.out = 100),
  return_label = runif(100)
)
# Create the returnPrediction object
rp <- create_return_prediction(data, "return_label")
# Check the initial setup
print(rp)

}
