% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seperation.R
\name{data_seperation}
\alias{data_seperation}
\title{Title}
\usage{
data_seperation(data, labels, features, start_date, end_date, seperation_date)
}
\arguments{
\item{data}{xyy}

\item{labels}{xy}

\item{features}{xy}

\item{start_date}{xy}

\item{end_date}{xy}

\item{seperation_date}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
