% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_pred_functions.R
\name{enet_pred}
\alias{enet_pred}
\title{Elastic Net Prediction Function}
\usage{
enet_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{A data frame with \code{stock_id}, \code{date}, \code{return_label}, and features.}

\item{test_data}{A data frame with \code{stock_id}, \code{date}, and features.}

\item{config}{A list of Elastic Net parameters (\code{alpha} and \code{lambda}).}
}
\value{
A tibble with \code{stock_id}, \code{date}, and \code{pred_return} matching the test data.
}
\description{
This function trains an Elastic Net model using the glmnet package and predicts outcomes
for the provided test data set based on the trained model.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c("stock_id", "date", "R1M_Usd", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
test_data_ex <- data_ml[101:150, c("stock_id", "date", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
enet_pred(train_data_ex, test_data_ex, config = list(alpha = 0.5, lambda = 0.1))
}
}
