% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_functions.R
\name{ensemble_weights}
\alias{ensemble_weights}
\title{Create Weights Based on Ensemble Methods}
\usage{
ensemble_weights(return_predictions, errors, method = "simple_average")
}
\arguments{
\item{return_predictions}{Tibble containing \code{stock_id}, \code{date}, and various prediction columns.}

\item{errors}{Tibble containing \code{stock_id}, \code{date}, and various error columns.}

\item{method}{Character. Method to use for ensemble weights. Possible values are:
\itemize{
\item "simple_average": Simple average of predictions.
\item "weighted_average": Weighted average based on the inverse of average errors.
\item "error_covariance": Using the covariance of errors to calculate weights.
}}
}
\value{
Numeric vector of ensemble predictions.
}
\description{
Create Weights Based on Ensemble Methods
}
