% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_config}
\alias{ensure_config}
\title{Ensure Configuration}
\usage{
ensure_config(config, default_params)
}
\arguments{
\item{config}{A list of configuration parameters.}

\item{default_params}{A list of default parameters.}
}
\value{
A list of configuration parameters with any missing arguments filled in with default values.
}
\description{
This function checks a configuration list for missing arguments and adds default values if necessary.
}
