% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_weight_func.R
\name{equal_weights_func}
\alias{equal_weights_func}
\title{Title}
\usage{
equal_weights_func(train_data, test_data, features = NULL)
}
\arguments{
\item{train_data}{xyx}

\item{test_data}{xy}

\item{features}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
