% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_largest_number}
\alias{find_largest_number}
\title{Helper function to find largest number after a certain string}
\usage{
find_largest_number(strings, model_name)
}
\arguments{
\item{strings}{vector of strings to search for}

\item{model_name}{string to search for the largest number after}
}
\value{
The largest number found after the model_name in the strings
}
\description{
Helper function to find largest number after a certain string
}
