% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestigatoR-package.R
\docType{package}
\name{InvestigatoR-package}
\alias{InvestigatoR}
\alias{InvestigatoR-package}
\title{InvestigatoR: Machine Learning for Factor Investing with R}
\description{
The package provides a front-end to backtesting investment portfolios using various machine learning and weighting algorithms. It is based on various features from the book Machine Learning for Factor Investing at \url{https://www.mlfactor.com/} by Guillaume Coqueret.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sstoeckl/InvestigatoR}
  \item Report bugs at \url{https://github.com/sstoeckl/InvestigatoR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Stoeckl \email{sebastian.stoeckl@uni.li} (\href{https://orcid.org/0000-0002-4196-6093}{ORCID}) (Package maintainer and co-author.)

Authors:
\itemize{
  \item Eric Schumann
  \item Andreas Pischetsrieder
  \item Christian Mahrun
  \item Noah Veltri
  \item Nicolas Baumann
  \item Aziza Ouali
  \item Nazim Yigitoglu
  \item Mikhail Zhabin
}

}
\keyword{internal}
