% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{leverage_metric}
\alias{leverage_metric}
\title{This function calculates and reports the leverage of the portfolio during training.
Leverage is defined as the sum of absolute portfolio weights for each time period.}
\usage{
leverage_metric()
}
\value{
A scalar leverage value for the current batch of data.
}
\description{
This function calculates and reports the leverage of the portfolio during training.
Leverage is defined as the sum of absolute portfolio weights for each time period.
}
