% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{MemmelSharpeTest}
\alias{MemmelSharpeTest}
\title{Check for Missing Values
Memmel-Corrected Sharpe Ratio Test}
\usage{
MemmelSharpeTest(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{A numeric vector or xts object representing the returns of portfolio x.}

\item{y}{A numeric vector or xts object representing the returns of portfolio y (benchmark).}

\item{alternative}{A string specifying the alternative hypothesis:
"two.sided" (default), "greater" (Ha: Sx > Sy), or "less" (Ha: Sx < Sy).}
}
\value{
A list containing the test statistic (\code{f}), p-value (\code{p}),
Sharpe ratio of portfolio x (\code{SRx}), Sharpe ratio of portfolio y (\code{SRy}),
and the alternative hypothesis (\code{alternative}).
}
\description{
Tests Sharpe ratios (SR) of two portfolios for significant differences,
based on the Jobson and Korkie (1981) method with Memmel's (2003) correction.
}
\examples{
# Example usage
x <- rnorm(100, mean = 0.01, sd = 0.02)
y <- rnorm(100, mean = 0.005, sd = 0.02)
result <- MemmelSharpeTest(x, y, alternative = "greater")
print(result)

}
