% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_data.R
\docType{data}
\name{my_data}
\alias{my_data}
\title{My simulated data}
\format{
A data frame with 100 rows
\describe{
\item{x}{Random Uniform Variable}
\item{y}{Random Uniform Variable}
}
}
\usage{
my_data
}
\description{
A simulated dataset that I am saving in my package.
}
\keyword{datasets}
