% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_pred_functions.R
\name{nnet_pred}
\alias{nnet_pred}
\title{Neural Network (nnet) Prediction Function}
\usage{
nnet_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{A data frame with \code{stock_id}, \code{date}, \code{return_label}, and features.}

\item{test_data}{A data frame with \code{stock_id}, \code{date}, and features.}

\item{config}{A list containing parameters for the neural network, such as \code{size} and \code{decay}.}
}
\value{
A tibble with \code{stock_id}, \code{date}, and \code{pred_return} matching the test data.
}
\description{
This function trains a neural network using the nnet package and makes predictions
for the test data set.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c("stock_id", "date", "R1M_Usd", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
test_data_ex <- data_ml[101:150, c("stock_id", "date", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
nnet_pred(train_data_ex, test_data_ex, config = list(size = 10, decay = 0.01))
}
}
