% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols_predictor.R
\name{ols_predictor}
\alias{ols_predictor}
\title{Title}
\usage{
ols_predictor(data, labels, features)
}
\arguments{
\item{data}{xy}

\item{labels}{xy}

\item{features}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
