% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{perf_met}
\alias{perf_met}
\title{Helper function to create performance metrics for a given set of returns, weights, and actual data}
\usage{
perf_met(returns_data, weights_data, actual_returns)
}
\arguments{
\item{returns_data}{return data}

\item{weights_data}{weights data}

\item{actual_returns}{actual realized returns}
}
\value{
A tibble with columns portfolio, mean, sd, SR, VaR_5, turnover, and hit_ratio
}
\description{
Helper function to create performance metrics for a given set of returns, weights, and actual data
}
