% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{plot.portfolio}
\alias{plot.portfolio}
\title{Plotting method for portfolioReturns S3 objects}
\usage{
\method{plot}{portfolio}(portfolio_object, type = NULL)
}
\arguments{
\item{portfolio_object}{A portfolioReturns S3 object}

\item{type}{default=NULL (standrad evaluation), Alternative: tq_performance_functions}
}
\value{
A ggplot object
}
\description{
Plotting method for portfolioReturns S3 objects
}
