% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portf_map.R
\name{portf_map}
\alias{portf_map}
\title{Title}
\usage{
portf_map(t, data, ticks, t_oos, m_offset, train_size, weight_func)
}
\arguments{
\item{t}{xy}

\item{data}{xy}

\item{ticks}{xy}

\item{t_oos}{xy}

\item{m_offset}{xy}

\item{train_size}{xy}

\item{weight_func}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
