% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neural_network_predictor.R
\name{predict_stock}
\alias{predict_stock}
\title{Title}
\usage{
predict_stock(
  train_features,
  train_labels,
  test_features,
  test_labels,
  epochs = 50,
  batches = 32
)
}
\arguments{
\item{train_features}{xy}

\item{train_labels}{xy}

\item{test_features}{xy}

\item{test_labels}{xy}

\item{epochs}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
