% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_predictor.R
\name{random_foreest_predictor}
\alias{random_foreest_predictor}
\title{Title}
\usage{
random_foreest_predictor(y, features, data, ntree)
}
\arguments{
\item{y}{xy}

\item{features}{xy}

\item{data}{xy}

\item{ntree}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy

}
