% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_pred_functions.R
\name{rf_pred}
\alias{rf_pred}
\title{Random Forest Function for Return Prediction}
\usage{
rf_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{A data frame with \code{stock_id}, \code{date}, \code{return_label}, and features.}

\item{test_data}{A data frame with \code{stock_id}, \code{date}, and features.}

\item{config}{A list of Random Forest configuration parameters (e.g., \code{num.trees}, \code{mtry}).}
}
\value{
A tibble with \code{stock_id}, \code{date}, and \code{pred_return} matching the test data.
}
\description{
This function trains a Random Forest model to predict returns based on the provided training data
and predicts returns for the test data.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c("stock_id", "date", "R1M_Usd", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
test_data_ex <- data_ml[101:150, c("stock_id", "date", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
rf_pred(train_data_ex, test_data_ex, config = list(num.trees = 200, mtry = 4))
}
}
