% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{select_dates_by_offset}
\alias{select_dates_by_offset}
\title{Helper function to create a tibble of training/prediction start and end dates based on a rolling window with an offset}
\usage{
select_dates_by_offset(dates, window_size, step_size, offset, rolling = TRUE)
}
\arguments{
\item{dates}{A vector of unique dates in increasing order}

\item{window_size}{A string that contains the window size, e.g. "3 days", "1 week", "2 months", "1 year"}

\item{step_size}{A string that contains the step size, e.g. "3 days", "1 week", "2 months", "1 year"}

\item{offset}{A string that contains the offset, e.g. "3 days", "1 week", "2 months", "1 year"}

\item{rolling}{A boolean that indicates whether the window should be rolling or expanding (FALSE)}
}
\value{
A tibble with columns training_start, training_end, prediction_start, prediction_end, and prediction_phase
}
\description{
Helper function to create a tibble of training/prediction start and end dates based on a rolling window with an offset
}
\examples{
dates <- seq(as.Date("2020-01-01"), as.Date("2023-01-01"), by = "month")
window_size <- "1 year"
step_size <- "3 months"
offset <- "1 month"
date_intervals <- select_dates_by_offset(dates, window_size, step_size, offset)
print(date_intervals)
date_intervals_exp <- select_dates_by_offset(dates, window_size, step_size, offset, rolling=FALSE)
print(date_intervals_exp)

}
