% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{summary.correlation}
\alias{summary.correlation}
\title{Summarize Portfolio Correlations}
\usage{
\method{summary}{correlation}(portfolio_object, benchmark_data = NULL, print = FALSE)
}
\arguments{
\item{portfolio_object}{A \code{portfolioReturns} object.}

\item{benchmark_data}{(Optional) A data frame with columns \code{date} and \code{return} representing an external benchmark.}

\item{print}{Logical. If \code{TRUE}, prints the correlation table using \code{kable}.}
}
\value{
A correlation matrix between portfolio returns and the benchmark.
}
\description{
Computes and optionally prints a correlation table between portfolio returns,
with the benchmark being the first column.
}
