% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{summary.portfolioReturns}
\alias{summary.portfolioReturns}
\title{Define the summary method for 'portfolioReturns' class}
\usage{
\method{summary}{portfolioReturns}(portfolio_object, type = NULL)
}
\arguments{
\item{portfolio_object}{A portfolioReturns object}

\item{type}{default=NULL (standard evaluation), Alternative: tq_performance_functions}
}
\value{
A summary satistics of the portfolio returns
}
\description{
Define the summary method for 'portfolioReturns' class
}
