% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{summary.returnPrediction}
\alias{summary.returnPrediction}
\title{Summary method for returnPrediction S3 object}
\usage{
\method{summary}{returnPrediction}(return_prediction_object, benchmark = NULL)
}
\arguments{
\item{benchmark}{Actual returns to compare against (not yet implemented)}

\item{object}{returnPrediction object}
}
\value{
Data frame with error statistics for each model
}
\description{
Summary method for returnPrediction S3 object
}
