% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_pred_functions.R
\name{svm_pred}
\alias{svm_pred}
\title{Support Vector Machine (SVM) Prediction Function}
\usage{
svm_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{A data frame with \code{stock_id}, \code{date}, \code{return_label}, and features.}

\item{test_data}{A data frame with \code{stock_id}, \code{date}, and features.}

\item{config}{A list of SVM configuration parameters (e.g., \code{kernel}, \code{type}, \code{gamma}, \code{cost}, \code{epsilon}).}
}
\value{
A tibble with \code{stock_id}, \code{date}, and \code{pred_return} matching the test data.
}
\description{
This function trains a Support Vector Machine (SVM) model to predict returns based on the provided
training data and makes predictions for the test data.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c("stock_id", "date", "R1M_Usd", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
test_data_ex <- data_ml[101:150, c("stock_id", "date", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
svm_pred(train_data_ex, test_data_ex, config = list(kernel = "linear", cost = 1))
}
}
