% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{weightpred_map}
\alias{weightpred_map}
\title{Weight Prediction Mapping Function (Helper)}
\usage{
weightpred_map(t, data_subset, indices, model_function, model_config)
}
\arguments{
\item{t}{Integer. Index of the current iteration.}

\item{data_subset}{Tibble. Subset of the data containing \code{stock_id}, \code{date}, \code{return_label}, features, and any optional benchmark/masking data.}

\item{indices}{Tibble. Contains \code{training_start}, \code{training_end}, \code{prediction_start}, and \code{prediction_end} dates.}

\item{model_function}{Character. Name of the weight prediction function to be invoked.}

\item{model_config}{List. Configuration parameters for the model function.}
}
\value{
Tibble containing \code{stock_id}, \code{date}, and predicted \code{weights}.
}
\description{
This helper function is used within \code{backtesting_weights} to map over time indices and generate weight predictions using specified ML models.
}
