% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_xgb.R
\name{weights_xgb}
\alias{weights_xgb}
\title{Title}
\usage{
weights_xgb(train_data, test_data, features)
}
\arguments{
\item{train_data}{xy}

\item{test_data}{xy}

\item{features}{xy}
}
\value{
xy
}
\description{
Title
}
\examples{
xy
}
