% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{xgb_pred}
\alias{xgb_pred}
\title{xgb function for return prediction}
\usage{
xgb_pred(train_data, test_data, config)
}
\arguments{
\item{train_data}{data frame with stock_id, date, return_label and features}

\item{test_data}{data frame with stock_id, date and features}

\item{config}{empty list, as ols does not need any configuration}

\item{fast}{logical, if TRUE, use fastLm from RcppArmadillo, else use lm from base R}
}
\value{
tibble with stock_id, date and pred_return matching the test_data
}
\description{
xgb function for return prediction
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100,c(1,2,96,5:10)]
test_data_ex <- data_ml[101:150,c(1,2,5:10)]
ols_pred(train_data_ex, test_data_ex, config=list(), fast=TRUE)
ols_pred(train_data_ex, test_data_ex, config=list(), fast=FALSE)
}
}
