% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{add_weight_model}
\alias{add_weight_model}
\title{Add a Weight Model to a portfolioReturns Object}
\usage{
add_weight_model(
  portfolio_object,
  model_name,
  new_weights,
  config,
  postprocessing_config = NULL
)
}
\arguments{
\item{portfolio_object}{An existing \code{portfolioReturns} object.}

\item{model_name}{A string specifying the name of the new weight model.}

\item{new_weights}{A tibble or data frame containing the new weights with columns \code{date}, \code{stock_id}, and \code{weight}.}
}
\value{
An updated \code{portfolioReturns} object with the new weight model added.
}
\description{
Adds a new weight model to an existing \code{portfolioReturns} object. Handles both benchmark and non-benchmark portfolios by adjusting weights and updating portfolio returns accordingly.
}
\examples{
\dontrun{
# Adding a weight model to a non-benchmark portfolio
portfolio_obj <- add_weight_model(
  portfolio_object = portfolio_obj,
  model_name = "Model_A",
  new_weights = new_weights_df
)

# Adding a weight model to a benchmark portfolio
portfolio_obj_bm <- add_weight_model(
  portfolio_object = portfolio_obj_bm,
  model_name = "Model_B",
  new_weights = new_weights_bm_df,
  benchmark_weights = benchmark_weights_df
)
}

}
