% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_objects.R
\name{create_portfolios}
\alias{create_portfolios}
\title{Create portfolioReturns S3 object}
\usage{
create_portfolios(data, label)
}
\arguments{
\item{data}{Data frame containing stock_id, date, and the specified return label.}

\item{label}{Column name of the actual return.}
}
\value{
A portfolioReturns S3 object.
}
\description{
Create portfolioReturns S3 object
}
\examples{
data <- tibble::tibble(
  stock_id = 1:100,
  date = rep(Sys.Date(), 100),
  return_label = runif(100)
)
pf <- create_portfolios(data, "return_label")
print(pf)
}
