% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{diversification_metric}
\alias{diversification_metric}
\title{This function calculates and reports the diversification of the portfolio during training,
using the Herfindahl-Hirschman Index (HHI), which is the sum of the squared portfolio weights.}
\usage{
diversification_metric()
}
\value{
A scalar diversification value (HHI) for the current batch of data.
}
\description{
This function calculates and reports the diversification of the portfolio during training,
using the Herfindahl-Hirschman Index (HHI), which is the sum of the squared portfolio weights.
}
