% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{dummy_mse_loss}
\alias{dummy_mse_loss}
\title{This function calculates the Mean Squared Error (MSE) between the actual returns (from column 4 of y_true)
and the predicted values (y_pred). It is used as a dummy function for debugging the custom loss function
input structure.}
\usage{
dummy_mse_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{A matrix where the 4th column contains the actual returns.}

\item{y_pred}{A vector of predicted portfolio weights/returns.}
}
\value{
The calculated MSE loss value.
}
\description{
This function calculates the Mean Squared Error (MSE) between the actual returns (from column 4 of y_true)
and the predicted values (y_pred). It is used as a dummy function for debugging the custom loss function
input structure.
}
