% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{enet_pred}
\alias{enet_pred}
\title{Elastic Net Prediction Function}
\usage{
enet_pred(train_data, test_data, config = list(), seed = NULL)
}
\arguments{
\item{train_data}{data frame with stock_id, date, return_label, and features}

\item{test_data}{data frame with stock_id, date, and features}

\item{config}{list with Elastic Net parameters (alpha and lambda)}

\item{seed}{integer. Seed for random number generation}
}
\value{
tibble with stock_id, date, and pred_return matching the test_data
}
\description{
This function trains an Elastic Net model using the glmnet package and predicts outcomes
for the provided test data set based on the trained model.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c(1,2,96,5:10)]
test_data_ex <- data_ml[101:150, c(1,2,5:10)]
enet_pred(train_data_ex, test_data_ex)
}
}
