% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_functions.R
\name{ensemble_weights}
\alias{ensemble_weights}
\title{Create weights based on various predictions, potentially using their errors with given constraints}
\usage{
ensemble_weights(return_predictions, errors, method = "simple_average")
}
\arguments{
\item{return_predictions}{tibble of stock_id, date, and various predictions}

\item{errors}{tibble of stock_id, date, and various prediction errors}

\item{method}{method to use for ensemble weights. Possible values are:
\itemize{
\item simple_average: simple average of predictions
\item weighted_average: weighted average based on the inverse of average errors
\item error_covariance: using the covariance of errors to calculate weights
}}
}
\value{
tibble with stock_id, date, and weight
}
\description{
Create weights based on various predictions, potentially using their errors with given constraints
}
\examples{
\dontrun{
}

}
