% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestigatoR-package.R
\docType{package}
\name{InvestigatoR-package}
\alias{InvestigatoR}
\alias{InvestigatoR-package}
\title{InvestigatoR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ericschumann12/InvestigatoR}
  \item Report bugs at \url{https://github.com/ericschumann12/InvestigatoR/issues}
}

}
\author{
\strong{Maintainer}: Eric Schumann \email{eric.schumann@uni.li}

Authors:
\itemize{
  \item Christian Mahrun \email{christian.mahrun@uni.li}
}

Other contributors:
\itemize{
  \item  [contributor]
}

}
\keyword{internal}
