% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{nnet_pred}
\alias{nnet_pred}
\title{Neural Network (nnet) Prediction Function}
\usage{
nnet_pred(train_data, test_data, config = list(), seed = NULL)
}
\arguments{
\item{train_data}{A data frame with stock_id, date, return_label, and features.}

\item{test_data}{A data frame with stock_id, date, and features.}

\item{config}{A list containing parameters for the neural network, such as size and decay.}

\item{seed}{integer. Seed for random number generation}
}
\value{
A tibble with stock_id, date, and predicted returns matching the test data.
}
\description{
This function trains a neural network using the nnet package and makes predictions
for the test data set.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, ]
test_data_ex <- data_ml[101:150, ]
nnet_pred(train_data_ex, test_data_ex, config = list())
}
}
