% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{ols_pred}
\alias{ols_pred}
\title{OLS Prediction Function}
\usage{
ols_pred(train_data, test_data, config = list(), fast = TRUE)
}
\arguments{
\item{train_data}{Data frame containing \code{stock_id}, \code{date}, \code{return_label}, and feature columns.}

\item{test_data}{Data frame containing \code{stock_id}, \code{date}, and feature columns.}

\item{config}{List. Configuration parameters for the model (not used in OLS).}

\item{fast}{Logical. If \code{TRUE}, use \code{fastLm} from \code{RcppArmadillo}; otherwise, use base \code{lm}.}
}
\value{
Tibble with \code{stock_id}, \code{date}, and \code{pred_return} matching the \code{test_data}.
}
\description{
OLS Prediction Function
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c("stock_id", "date", "R1M_Usd", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
test_data_ex <- data_ml[101:150, c("stock_id", "date", "Div_Yld", "Eps", "Mkt_Cap_12M_Usd", "Mom_11M_Usd", "Ocf", "Pb", "Vol1Y_Usd")]
ols_pred(train_data_ex, test_data_ex, config = list(), fast = TRUE)
}
}
