% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappers.R
\name{retpred_map}
\alias{retpred_map}
\title{Return Prediction Mapping Function (Helper)}
\usage{
retpred_map(t, data_subset, indices, model_function, model_config)
}
\arguments{
\item{t}{integer, index of the current iteration}

\item{data_subset}{tibble, subset of the data (stock_id, date, return_label, features)}

\item{indices}{list, list of training and prediction indices}

\item{model_function}{character, name of the prediction (model) function}

\item{model_config}{list, configuration for the model}
}
\value{
tibble that contains stock id, date, and predicted return
}
\description{
Return Prediction Mapping Function (Helper)
}
