% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{svm_pred}
\alias{svm_pred}
\title{Support Vector Machine (SVM) Prediction Function}
\usage{
svm_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{Data frame containing training data with features and a target variable.}

\item{test_data}{Data frame containing test data with features.}

\item{config}{List containing any additional parameters for the SVM model.}
}
\value{
A tibble with stock_id, date, and predicted returns.
}
\description{
Trains a Support Vector Machine using the e1071 package and predicts outcomes
for the provided test data set based on the trained model.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100,c(1,2,96,5:10)]
test_data_ex <- data_ml[101:150,c(1,2,5:10)]
svm_pred(train_data_ex, test_data_ex, config)
}
}
