% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPredictors.R
\name{svm_pred}
\alias{svm_pred}
\title{Support Vector Machine (SVM) Prediction Function}
\usage{
svm_pred(train_data, test_data, config = list())
}
\arguments{
\item{train_data}{A data frame with stock_id, date, return_label, and features.}

\item{test_data}{A data frame with stock_id, date, and features.}

\item{config}{A list of SVM configuration parameters (e.g., kernel, type, gamma, cost, epsilon).}
}
\value{
A tibble with stock_id, date, and predicted returns matching the test data.
}
\description{
This function trains a Support Vector Machine (SVM) model to predict returns based on the provided
training data and makes predictions for the test data.
}
\examples{
\dontrun{
data(data_ml)
train_data_ex <- data_ml[1:100, c(1,2,96,5:10)]
test_data_ex <- data_ml[101:150, c(1,2,5:10)]
svm_pred(train_data_ex, test_data_ex, config = list())
}
}
