% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred_functions.R
\name{turnover_metric}
\alias{turnover_metric}
\title{This function calculates and reports the portfolio turnover during training.}
\usage{
turnover_metric()
}
\value{
A scalar turnover value for the current batch of data.
}
\description{
This function calculates and reports the portfolio turnover during training.
}
